Added auto increment for PK

+++ pharmacy-management-system/src/main/resources/db/migration/v1.6__create_table_producatori.sql
@@ -1,6 +1,6 @@
 -- Crearea tabelului PRODUCATORI
 CREATE TABLE producatori (
-                             id_producator DECIMAL(5, 0),
+                             id_producator INT,
                              CONSTRAINT PROD_id_prod_PK PRIMARY KEY(id_producator),
                              producator VARCHAR(50) NOT NULL,
                              CONSTRAINT PROD_prod_UK UNIQUE(producator),
